{"schema":"decodex/commit/1","summary":"Split CLI and tracker modules","authority":"manual"}
