+++ apps/decodex/src/cli/attempt_command.rs
@@ -1,0 +1,113 @@
+use std::{
+	fs,
+	io::{self, Read as _},
+};
+
+use clap::Args;
+use serde::{Deserialize, Serialize};
+
+use crate::{
+	cli::ProjectConfigArgs,
+	orchestrator::{self, IssueDispatchMode, RunOnceRequest},
+	prelude::{Result, eyre},
+};
+
+#[derive(Clone, Debug, Eq, PartialEq, Deserialize, Serialize)]
+pub(crate) struct AttemptRequest {
+	#[serde(default)]
+	pub(crate) dry_run: bool,
+	pub(crate) issue_id: String,
+	pub(crate) issue_state: String,
+	pub(crate) initial_issue_state: Option<String>,
+	#[serde(default)]
+	pub(crate) lease_preacquired: bool,
+	pub(crate) issue_claim_fd: Option<i32>,
+	pub(crate) dispatch_slot_fd: Option<i32>,
+	pub(crate) dispatch_slot_index: Option<usize>,
+	pub(crate) dispatch_mode: AttemptDispatchMode,
+	pub(crate) run_id: String,
+	pub(crate) attempt_number: i64,
+	pub(crate) retry_budget_base: i64,
+	pub(crate) workflow_snapshot: String,
+}
+
+#[derive(Debug, Args)]
+pub(in crate::cli) struct AttemptCommand {
+	#[command(flatten)]
+	pub(in crate::cli) project_config: ProjectConfigArgs,
+	/// Structured request file path, or `-` to read the request from stdin.
+	#[arg(value_name = "REQUEST", default_value = "-")]
+	pub(in crate::cli) request: String,
+}
+impl AttemptCommand {
+	pub(in crate::cli) fn run(&self) -> Result<()> {
+		let request = read_attempt_request(&self.request)?;
+
+		orchestrator::run_once(RunOnceRequest {
+			config_path: self.project_config.as_path(),
+			dry_run: request.dry_run,
+			explain_queue: false,
+			preferred_issue_id: Some(request.issue_id.as_str()),
+			preferred_issue_state: Some(request.issue_state.as_str()),
+			preferred_initial_issue_state: request.initial_issue_state.as_deref(),
+			preferred_lease_acquired: request.lease_preacquired,
+			preferred_issue_claim_fd: request.issue_claim_fd,
+			preferred_dispatch_slot_fd: request.dispatch_slot_fd,
+			preferred_dispatch_slot_index: request.dispatch_slot_index,
+			preferred_dispatch_mode: Some(request.dispatch_mode.into()),
+			preferred_run_id: Some(request.run_id.as_str()),
+			preferred_attempt_number: Some(request.attempt_number),
+			preferred_retry_budget_base: Some(request.retry_budget_base),
+			preferred_workflow_snapshot: Some(request.workflow_snapshot.as_str()),
+		})
+	}
+}
+
+impl From<AttemptDispatchMode> for IssueDispatchMode {
+	fn from(value: AttemptDispatchMode) -> Self {
+		match value {
+			AttemptDispatchMode::Normal => Self::Normal,
+			AttemptDispatchMode::Program => Self::Program,
+			AttemptDispatchMode::Retry => Self::Retry,
+			AttemptDispatchMode::ReviewRepair => Self::ReviewRepair,
+			AttemptDispatchMode::Closeout => Self::Closeout,
+		}
+	}
+}
+
+#[derive(Clone, Copy, Debug, Eq, PartialEq, Deserialize, Serialize)]
+#[serde(rename_all = "kebab-case")]
+pub(crate) enum AttemptDispatchMode {
+	Normal,
+	Program,
+	Retry,
+	ReviewRepair,
+	Closeout,
+}
+impl From<IssueDispatchMode> for AttemptDispatchMode {
+	fn from(value: IssueDispatchMode) -> Self {
+		match value {
+			IssueDispatchMode::Normal => Self::Normal,
+			IssueDispatchMode::Program => Self::Program,
+			IssueDispatchMode::Retry => Self::Retry,
+			IssueDispatchMode::ReviewRepair => Self::ReviewRepair,
+			IssueDispatchMode::Closeout => Self::Closeout,
+		}
+	}
+}
+
+fn read_attempt_request(request: &str) -> Result<AttemptRequest> {
+	let raw = if request == "-" {
+		let mut raw = String::new();
+
+		io::stdin().read_to_string(&mut raw)?;
+
+		raw
+	} else {
+		fs::read_to_string(request)?
+	};
+
+	serde_json::from_str(&raw).map_err(|error| {
+		eyre::eyre!("Failed to parse `_attempt` request from `{}`: {error}", request)
+	})
+}
